requirements.txt
groq
pyperclip
keyboard
pywin32
customtkinter
pystray
Pillow
python-dotenv